move test for writeProblem to separate file
